Move port-collecting bits to own file.
